+++ R/brightnessgraph.R
@@ -14,7 +14,7 @@
 #' @noRd
 bg_reduction <- function(set, edo=12, rounder=10) {
   card <- length(set)
-  scalar_interval_matrix <- sim(set,edo)
+  scalar_interval_matrix <- sim(set,edo=edo)
   sums <- colSums(scalar_interval_matrix)
 
   comparisons <- -1*brightness_comparisons(set, edo, rounder)
@@ -88,12 +88,12 @@ bg_reduction <- function(set, edo=12, rounder=10) {
 #' werckmeister_3 <- z(werck_ratios)
 #' brightnessgraph(werckmeister_3, show_sums=FALSE, show_pitches=FALSE)
 #' 
-#' @returns `NULL` and plots a brightness graph in the graphic device
+#' @returns Invisibly, an igraph graph object (the structure of the plotted brightness graph)
 #' @export
 brightnessgraph <- function(set, numdigits=2, show_sums=TRUE, show_pitches=TRUE, fixed_do=FALSE,
                             edo=12, rounder=10) {
   card <- length(set)
-  scalar_interval_matrix <- sim(set,edo)
+  scalar_interval_matrix <- sim(set,edo=edo)
   sums <- colSums(scalar_interval_matrix)
   y_coords <- sums
 
@@ -140,7 +140,7 @@ brightnessgraph <- function(set, numdigits=2, show_sums=TRUE, show_pitches=TRUE,
   if (fixed_do==TRUE) {
     pitch_labels <- sapply(0:(card-1), rotate, x=set, edo=edo)
   } else {
-    pitch_labels <- sim(set,edo=edo)
+    pitch_labels <- sim(set, edo=edo)
   }
   pitch_labels <- apply(apply(pitch_labels,2, round, digits=numdigits), 2, paste, collapse=", ")
 
@@ -165,9 +165,9 @@ brightnessgraph <- function(set, numdigits=2, show_sums=TRUE, show_pitches=TRUE,
 
   if (class(label_matrix)[1]=="character") label_matrix <- as.matrix(label_matrix)
 
-  label_vector <- apply(label_matrix,1,paste,collapse="")
+  label_vector <- apply(label_matrix, 1, paste, collapse="")
 
   bg <- igraph::graph_from_adjacency_matrix(reduced_comparisons)
-  plot(bg,layout=layout_matrix, vertex.shape="none", vertex.label=label_vector)
-  invisible()
+  plot(bg, layout=layout_matrix, vertex.shape="none", vertex.label=label_vector)
+  invisible(bg)
 }